Remove Keras translation
Over the recent years Keras has turned from a thin convenience
API into a monster that drags along the whole Tensorflow ecosystem.
This, in turn, only works on modern computers - lack of AVX instructions
crashes the process on the first import of Keras on such a machine.

The existing proof of concept translation to Theano was not too advanced
nor practically useful in any case and now it is a liability that makes
it hard to support.